pyproject.toml
[project]
name = "idiograph"
version = "0.1.0"
description = "Semantic graph system for VFX and AI workflows"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.86.0",
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "networkx>=3.6.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "typer>=0.12.0",
]

[project.scripts]
idiograph = "idiograph.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/idiograph"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
